+++ src/Mapping/ExtensionMetadataFactory.php
@@ -72,18 +72,30 @@ class ExtensionMetadataFactory
 
     /**
      * @param Reader|AttributeReader|object|null $annotationReader
+     *
+     * @note Providing any object as the third argument is deprecated, as of 4.0 an {@see AttributeReader} will be required
      */
     public function __construct(ObjectManager $objectManager, string $extensionNamespace, ?object $annotationReader = null, ?CacheItemPoolInterface $cacheItemPool = null)
     {
-        if (null !== $annotationReader && !$annotationReader instanceof Reader && !$annotationReader instanceof AttributeReader) {
-            Deprecation::trigger(
-                'gedmo/doctrine-extensions',
-                'https://github.com/doctrine-extensions/DoctrineExtensions/pull/2258',
-                'Providing an annotation reader which does not implement %s or is not an instance of %s to %s is deprecated.',
-                Reader::class,
-                AttributeReader::class,
-                static::class
-            );
+        if (null !== $annotationReader) {
+            if ($annotationReader instanceof Reader) {
+                Deprecation::trigger(
+                    'gedmo/doctrine-extensions',
+                    'https://github.com/doctrine-extensions/DoctrineExtensions/pull/2772',
+                    'Annotations support is deprecated, migrate your application to use attributes and pass an instance of %s to the %s constructor instead.',
+                    AttributeReader::class,
+                    static::class
+                );
+            } elseif (!$annotationReader instanceof AttributeReader) {
+                Deprecation::trigger(
+                    'gedmo/doctrine-extensions',
+                    'https://github.com/doctrine-extensions/DoctrineExtensions/pull/2258',
+                    'Providing an annotation reader which does not implement %s or is not an instance of %s to %s is deprecated.',
+                    Reader::class,
+                    AttributeReader::class,
+                    static::class
+                );
+            }
         }
 
         $this->objectManager = $objectManager;
@@ -240,19 +252,19 @@ class ExtensionMetadataFactory
                 }
             }
 
-            if ($driver instanceof AnnotationDriverInterface) {
+            if ($driver instanceof AttributeDriverInterface) {
                 if (null === $this->annotationReader) {
                     throw new RuntimeException("Cannot use metadata driver ({$driverClassName}), an annotation or attribute reader was not provided.");
                 }
 
-                if ($driver instanceof AttributeDriverInterface) {
+                if ($driver instanceof AnnotationDriverInterface) {
+                    $driver->setAnnotationReader($this->annotationReader);
+                } else {
                     if ($this->annotationReader instanceof AttributeReader) {
                         $driver->setAnnotationReader($this->annotationReader);
                     } else {
                         $driver->setAnnotationReader(new AttributeAnnotationReader(new AttributeReader(), $this->annotationReader));
                     }
-                } else {
-                    $driver->setAnnotationReader($this->annotationReader);
                 }
             }
         }