+++ src/ReferenceIntegrity/Mapping/Driver/Attribute.php
@@ -9,17 +9,61 @@
 
 namespace Gedmo\ReferenceIntegrity\Mapping\Driver;
 
+use Gedmo\Exception\InvalidMappingException;
 use Gedmo\Mapping\Annotation\ReferenceIntegrity;
-use Gedmo\Mapping\Driver\AttributeDriverInterface;
+use Gedmo\Mapping\Driver\AbstractAnnotationDriver;
+use Gedmo\ReferenceIntegrity\Mapping\Validator;
 
 /**
- * This is an attribute mapping driver for ReferenceIntegrity
- * behavioral extension. Used for extraction of extended
- * metadata from attributes specifically for ReferenceIntegrity
- * extension.
+ * Mapping driver for the reference integrity extension which reads extended metadata from attributes on a class with referential integrity.
+ *
+ * @author Evert Harmeling <evert.harmeling@freshheads.com>
  *
  * @internal
  */
-final class Attribute extends Annotation implements AttributeDriverInterface
+class Attribute extends AbstractAnnotationDriver
 {
+    /**
+     * Mapping object for the reference integrity extension.
+     */
+    public const REFERENCE_INTEGRITY = ReferenceIntegrity::class;
+
+    /**
+     * Unimplemented mapping object for actions within the reference integrity extension.
+     *
+     * @deprecated since gedmo/doctrine-extensions 3.16, will be removed in version 4.0; actions are defined as {@see Validator} class constants instead.
+     */
+    public const ACTION = 'Gedmo\\Mapping\\Annotation\\ReferenceIntegrityAction';
+
+    public function readExtendedMetadata($meta, array &$config)
+    {
+        $validator = new Validator();
+        $reflClass = $this->getMetaReflectionClass($meta);
+
+        foreach ($reflClass->getProperties() as $reflProperty) {
+            if ($referenceIntegrity = $this->reader->getPropertyAnnotation($reflProperty, self::REFERENCE_INTEGRITY)) {
+                \assert($referenceIntegrity instanceof ReferenceIntegrity);
+
+                $property = $reflProperty->getName();
+
+                if (!$meta->hasField($property)) {
+                    throw new InvalidMappingException(sprintf('Unable to find reference integrity [%s] as mapped property in entity - %s', $property, $meta->getName()));
+                }
+
+                $fieldMapping = $meta->getFieldMapping($property);
+
+                if (!isset($fieldMapping['mappedBy'])) {
+                    throw new InvalidMappingException(sprintf("'mappedBy' should be set on '%s' in '%s'", $property, $meta->getName()));
+                }
+
+                if (!in_array($referenceIntegrity->value, $validator->getIntegrityActions(), true)) {
+                    throw new InvalidMappingException(sprintf('Field - [%s] does not have a valid integrity option, [%s] in class - %s', $property, implode(', ', $validator->getIntegrityActions()), $meta->getName()));
+                }
+
+                $config['referenceIntegrity'][$property] = $referenceIntegrity->value;
+            }
+        }
+
+        return $config;
+    }
 }